// frontend/generated/models/VideoDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 8lind8box Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { AccountDto } from './AccountDto';
import {
    AccountDtoFromJSON,
    AccountDtoFromJSONTyped,
    AccountDtoToJSON,
    AccountDtoToJSONTyped,
} from './AccountDto';

/**
 * 
 * @export
 * @interface VideoDto
 */
export interface VideoDto {
    /**
     * 
     * @type {number}
     * @memberof VideoDto
     */
    videoId?: number;
    /**
     * 
     * @type {AccountDto}
     * @memberof VideoDto
     */
    account?: AccountDto;
    /**
     * 
     * @type {number}
     * @memberof VideoDto
     */
    slotId?: number;
    /**
     * 
     * @type {string}
     * @memberof VideoDto
     */
    url?: string;
    /**
     * 
     * @type {string}
     * @memberof VideoDto
     */
    description?: string;
    /**
     * 
     * @type {boolean}
     * @memberof VideoDto
     */
    isVisible?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof VideoDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof VideoDto
     */
    updatedAt?: Date;
    /**
     * 
     * @type {boolean}
     * @memberof VideoDto
     */
    isVerified?: boolean;
}

/**
 * Check if a given object implements the VideoDto interface.
 */
export function instanceOfVideoDto(value: object): value is VideoDto {
    return true;
}

export function VideoDtoFromJSON(json: any): VideoDto {
    return VideoDtoFromJSONTyped(json, false);
}

export function VideoDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): VideoDto {
    if (json == null) {
        return json;
    }
    return {
        
        'videoId': json['videoId'] == null ? undefined : json['videoId'],
        'account': json['account'] == null ? undefined : AccountDtoFromJSON(json['account']),
        'slotId': json['slotId'] == null ? undefined : json['slotId'],
        'url': json['url'] == null ? undefined : json['url'],
        'description': json['description'] == null ? undefined : json['description'],
        'isVisible': json['isVisible'] == null ? undefined : json['isVisible'],
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
        'updatedAt': json['updatedAt'] == null ? undefined : (new Date(json['updatedAt'])),
        'isVerified': json['isVerified'] == null ? undefined : json['isVerified'],
    };
}

export function VideoDtoToJSON(json: any): VideoDto {
    return VideoDtoToJSONTyped(json, false);
}

export function VideoDtoToJSONTyped(value?: VideoDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'videoId': value['videoId'],
        'account': AccountDtoToJSON(value['account']),
        'slotId': value['slotId'],
        'url': value['url'],
        'description': value['description'],
        'isVisible': value['isVisible'],
        'createdAt': value['createdAt'] == null ? undefined : ((value['createdAt']).toISOString()),
        'updatedAt': value['updatedAt'] == null ? undefined : ((value['updatedAt']).toISOString()),
        'isVerified': value['isVerified'],
    };
}